+++ src/detect-engine-content-inspection.h
@@ -68,7 +68,7 @@ bool DetectEngineContentInspectionBuffer(DetectEngineCtx *de_ctx, DetectEngineTh
         const Signature *s, const SigMatchData *smd, Packet *p, Flow *f, const InspectionBuffer *b,
         const enum DetectContentInspectionType inspection_mode);
 
-uint8_t DetectEngineContentInspectionBufferMulti(DetectEngineCtx *de_ctx,
+int DetectEngineContentInspectionBufferMulti(DetectEngineCtx *de_ctx,
         DetectEngineThreadCtx *det_ctx, const Signature *s, const SigMatchData *smd, Flow *f,
         const InspectionBuffer *b, uint32_t local_id);
